Move handlers packages to internal
