+++ internal/handlers/handlers.go
@@ -9,16 +9,16 @@ import (
 	"strings"
 	"time"
 
-	"github.com/artifacthub/hub/cmd/hub/handlers/apikey"
-	"github.com/artifacthub/hub/cmd/hub/handlers/helpers"
-	"github.com/artifacthub/hub/cmd/hub/handlers/org"
-	"github.com/artifacthub/hub/cmd/hub/handlers/pkg"
-	"github.com/artifacthub/hub/cmd/hub/handlers/repo"
-	"github.com/artifacthub/hub/cmd/hub/handlers/static"
-	"github.com/artifacthub/hub/cmd/hub/handlers/stats"
-	"github.com/artifacthub/hub/cmd/hub/handlers/subscription"
-	"github.com/artifacthub/hub/cmd/hub/handlers/user"
-	"github.com/artifacthub/hub/cmd/hub/handlers/webhook"
+	"github.com/artifacthub/hub/internal/handlers/apikey"
+	"github.com/artifacthub/hub/internal/handlers/helpers"
+	"github.com/artifacthub/hub/internal/handlers/org"
+	"github.com/artifacthub/hub/internal/handlers/pkg"
+	"github.com/artifacthub/hub/internal/handlers/repo"
+	"github.com/artifacthub/hub/internal/handlers/static"
+	"github.com/artifacthub/hub/internal/handlers/stats"
+	"github.com/artifacthub/hub/internal/handlers/subscription"
+	"github.com/artifacthub/hub/internal/handlers/user"
+	"github.com/artifacthub/hub/internal/handlers/webhook"
 	"github.com/artifacthub/hub/internal/hub"
 	"github.com/artifacthub/hub/internal/img"
 	"github.com/go-chi/chi"
