+++ internal/handlers/static/testdata/image.png
