+++ internal/handlers/stats/handlers_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/artifacthub/hub/cmd/hub/handlers/helpers"
+	"github.com/artifacthub/hub/internal/handlers/helpers"
 	"github.com/artifacthub/hub/internal/stats"
 	"github.com/artifacthub/hub/internal/tests"
 	"github.com/rs/zerolog"
