+++ internal/handlers/user/handlers.go
@@ -16,7 +16,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/artifacthub/hub/cmd/hub/handlers/helpers"
+	"github.com/artifacthub/hub/internal/handlers/helpers"
 	"github.com/artifacthub/hub/internal/hub"
 	"github.com/artifacthub/hub/internal/user"
 	"github.com/coreos/go-oidc"
