+++ internal/handlers/webhook/handlers.go
@@ -7,7 +7,7 @@ import (
 	"net/http"
 	"text/template"
 
-	"github.com/artifacthub/hub/cmd/hub/handlers/helpers"
+	"github.com/artifacthub/hub/internal/handlers/helpers"
 	"github.com/artifacthub/hub/internal/hub"
 	"github.com/artifacthub/hub/internal/notification"
 	"github.com/go-chi/chi"
